+++ lib/sbb_design_system_mobile.dart
@@ -22,7 +22,7 @@ export 'src/segmented_button/sbb_segmented_button.dart';
 export 'src/select/sbb_select.dart';
 export 'src/shared/shared.dart';
 export 'src/slider/sbb_slider.dart';
-export 'src/status/sbb_status.dart';
+export 'src/status/status.dart';
 export 'src/stepper/stepper.dart';
 export 'src/switch/switch.dart';
 export 'src/tab_bar/tab_bar.dart';