+++ lib/src/theme/sbb_theme.dart
@@ -5,6 +5,7 @@ import 'package:sbb_design_system_mobile/src/checkbox/theme/default_sbb_checkbox
 import 'package:sbb_design_system_mobile/src/radio/theme/default_sbb_radio_theme_data.dart';
 
 import '../button/theme/default_button_themes.dart';
+import '../status/theme/default_sbb_status_theme.dart';
 
 const sbbDefaultSpacing = 16.0;
 
@@ -22,6 +23,7 @@ class SBBTheme {
     SBBControlStyles? controlStyles,
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBGroupStyle? groupStyle,
+    SBBStatusThemeData? statusTheme,
     SBBTextTheme? textTheme,
     SBBToastStyle? toastStyle,
   }) => createTheme(
@@ -36,6 +38,7 @@ class SBBTheme {
     controlStyles: controlStyles,
     headerBoxStyle: headerBoxStyle,
     groupStyle: groupStyle,
+    statusTheme: statusTheme,
     textTheme: textTheme,
     toastStyle: toastStyle,
   );
@@ -51,6 +54,7 @@ class SBBTheme {
     SBBControlStyles? controlStyles,
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBGroupStyle? groupStyle,
+    SBBStatusThemeData? statusTheme,
     SBBTextTheme? textTheme,
     SBBToastStyle? toastStyle,
   }) => createTheme(
@@ -65,6 +69,7 @@ class SBBTheme {
     controlStyles: controlStyles,
     headerBoxStyle: headerBoxStyle,
     groupStyle: groupStyle,
+    statusTheme: statusTheme,
     textTheme: textTheme,
     toastStyle: toastStyle,
   );
@@ -81,6 +86,7 @@ class SBBTheme {
     SBBControlStyles? controlStyles,
     SBBHeaderBoxStyle? headerBoxStyle,
     SBBGroupStyle? groupStyle,
+    SBBStatusThemeData? statusTheme,
     SBBTextTheme? textTheme,
     SBBToastStyle? toastStyle,
   }) {
@@ -112,6 +118,9 @@ class SBBTheme {
     final defaultGroupStyle = SBBGroupStyle.$default(baseStyle: mergedBaseStyle);
     final mergedGroupStyle = groupStyle.merge(defaultGroupStyle);
 
+    final defaultStatusTheme = DefaultSBBStatusTheme(baseStyle: mergedBaseStyle);
+    final mergedStatusTheme = defaultStatusTheme.merge(statusTheme);
+
     final defaultTextTheme = SBBTextTheme.$default(baseStyle: mergedBaseStyle);
     final mergedTextTheme = defaultTextTheme.merge(textTheme);
 
@@ -129,6 +138,7 @@ class SBBTheme {
       controlStyles: mergedControlStyles,
       headerBoxStyle: mergedHeaderBoxStyle,
       groupStyle: mergedGroupStyle,
+      statusTheme: mergedStatusTheme,
       textTheme: mergedTextTheme,
       toastStyle: mergedToastStyle,
     );
@@ -145,6 +155,7 @@ class SBBTheme {
     required SBBControlStyles controlStyles,
     required SBBHeaderBoxStyle headerBoxStyle,
     required SBBGroupStyle groupStyle,
+    required SBBStatusThemeData statusTheme,
     required SBBTextTheme textTheme,
     required SBBToastStyle toastStyle,
   }) {
@@ -177,6 +188,7 @@ class SBBTheme {
         controlStyles,
         headerBoxStyle,
         groupStyle,
+        statusTheme,
         textTheme,
         toastStyle,
       ],